Automatically generate a netrc file according to basic http auth in url
For every mirror url in a pinned json, try to extract netrc credentials. Then, write those
credentials into a file that every generated http_file for the various maven artifacts will use.

Alternative way of implementing would be by allowing you to pass a pre-made netrc and that can be
added in later if we should choose to do so, although it may be somewhat difficult to do so due to
there being at least three different repositories in play making it difficult to use raw paths.