Pre-pend HOME netrc to generated netrc for inline http auth
http_file will, by default, take in the $HOME/.netrc file to use for authentication.

This was broken by the earlier attempt to auto-generate netrc as a workaround for a lack of
inline http auth in http_file.

We now prepend the HOME netrc to the generated netrc file so that the HOME netrc serves as a
base which will get overlaid with the credentials in mirror_urls.