Add a `java_export` rule (#413)
This adds a set of rules that help make this happen:

  * `java_export` is a macro that is the user-facing entry point
  * `pom_file` for generating a pom file.
  * `javadoc` for generating javadocs from a given target
  * `maven_project_jar` for merging jars within a "project" together
  * `maven_publish` to actually push outputs to a maven repo

What is a project? Effectively every dependency that is not tagged
with `maven_coordinates=`, and every dependency of those that are not
tagged, and so on.

The generated pom file contains an entry for every first-order maven
dependency.

Included is a simple example, showing how to use
`java_export`. Running a `bazel query --output=build` will provide
some insight into how the macro expands.