Outdated: handle more maven-metadata.xml files
Fix bazelbuild/rules_jvm_external#507

The current `maven-metadata.xml` for `javax.inject:javax.inject` is missing a
release or latest tag.

    <metadata>
     <groupId>javax.inject</groupId>
     <artifactId>javax.inject</artifactId>
     <version>1</version>
     <versioning>
      <versions>
       <version>1</version>
      </versions>
      <lastUpdated>20100720032040</lastUpdated>
     </versioning>
    </metadata>

In this case, grab the first version in the list.

With the new code paths, the control flow was getting hard to track,
so I switched to returning at decision time.