Add a `deploy_env` attribute to `java_export`
This allows `java_export` to be used to generate things such as
plugins for Apache Spark or jar files for use in JEE servers by
excluding provided dependencies from the generated jars. This is
analagous to maven's `provided` scope, and maps to the `deploy_env`
attribute of `java_binary`.

The difference between `java_export` and `java_binary`'s `deploy_env`
is that `java_export` allows anything that provides a `JavaInfo` as an
input to `deploy_env`. This is because the expected use-cases are for
things that are normally provided as regular imports from a maven
dependency, rather than a custom binary.