Add analysis test to check jvm_import tags
This change adds an analysis test to verify the existence of the maven_url and
maven_coordinate tags in `jvm_import`, generated by maven_install. The analysis
test consumes an aspect which propagates the tags with the TagsInfo provider.

The test suite, containing the single analysis test, can be run with:
% bazel test tests/unit/jvm_import:jvm_import_tests