Merge jar files without unpacking them
It's possible to create jar files that do not unpack cleanly on
a file system. For example, they can include a _file_ named
`foo/bar` and another file called `foo`. Attempting to merge
jars by unpacking to disk and then repacking them will fail
because it's not possible to unpack cleanly.

This change does everything in-memory, without needing to create
intermediate temporary files.