Simplify and optimize manifest modification logic (#33)
* Simplify and optimize manifest modification logic

We only need to modify a file in the Jar but end up manually copying all
entries. For small jars this is not an issue, but for larger jars it has
terrible performance (specially in ci).
E.g on my M1 mac, stamping buckCompiler jar takes ~20 seconds with the
naive implementation. With the changes in this commit it now takes
~1.5s. The performance is even worse on ci which has terrible io (+10min
just to stamp this jar)

* avoid creating file with bad attrs