Allow the compile jars to strip illegal automatic module names
There are plenty of jars out there that have an `Automatic-Module-Name`
that isn't legal. Normally, this doesn't matter too much, but when
compiling code with a modern compiler bazel will use a module path,
and that cares _a lot_, causing builds to fail.

To rectify this, we can pass an additional flag when creating the
compile jar that will strip illegal automatic module names from the
generated manifests.