Add index of packages in jar files when pinning (#716)
This allows other tooling, e.g. gazelle, to make use of this
information. This can be useful for dependency inference, dependency
analysis, and suggesting missing dependencies.

This makes sense to do at pin time because the lifecycle of this
information is closely tied to the artifacts being fetched/pinned.
Generating the information is cheap enough that this shouldn't
significantly slow down pins, but is expensive enough that other tooling
may want to avoid doing so on the fly at runtime, so is useful to cache.