Add workspace_prefix_for_pinned_deps
When artifacts are pinned, `maven_install` would create individual
external repos for each dependency. The name of such an external repo is
simply the maven coordinate with all punctuation replaced by `_`. This
can cause name conflicts for code bases that are already using
`http_archive` to download thousands of dependencies through some other
mechanism but would like to migrate to use `maven_install`.

This PR tries to fix this by specifying a prefix for intermediate
external repos so that name collision becomes manageable.