Create a Maven BOM
In multi-project Maven repos, it's common to create a
Bill of Materials (BOM) `pom.xml` that shares the common
dependencies the project uses, pinned at specific versions.

The `maven_bom` rule gathers multiple `java_export` targets
into a single unit, and creates the BOM from them. The
dependencies included in the BOM are:

 - Ones that appear as dependencies of two or more of the
  `java_export`s the BOM is being created for.
 - Are not one of the `java_export`s the BOM is being
   created for.