Generate both a BOM and a dependency BOM (#750)
This changes how we generate BOMs to something that is
more often seen "in the wild". The main BOM now contains
just a list of all the `java_export` targets that are
listed in the `java_exports` parameter.

If the `dependencies_maven_coordinates` parameter is
set, then we shall also generate a "dependencies" BOM,
which lists all the first-order deps of the `java_exports`
as well as having a reference to the parent BOM.

Each of these are ultimately a `pom.xml` file, and each
of these is generated using its own rule.