Move `outdated` deps to regular rje deps
This means that it's in the lock file (so no-one needs to
re-resolve the dependencies), and also means that the tests
which are failing as we attempt to move to the BCR pass.

As part of this work, we'll need another RJE release, so
this also bumps the version in the module file.