Generate a stub so bzl_library works as expected (#761)
When we freeze deps (as we do in contrib_rules_jvm), we keep a copy of the generated workspace in a zip. This is then unpacked and used as a dependency, thus avoiding the need to do a normal initialisation process of "first make rje available, then load its deps, then call maven_install, then load the generated workspace". This allows maven_install to be used for deps that take a regular "have a function for dependencies, and a function for setting up the ruleset" structure.

However, when we do that, stardoc will attempt to index through the "frozen" workspace to find the definitions from rje. This PR ensures that behaviour works as expected.