fix NO_PROXY handling under Java 9+
Remove single-quoting of the `-Dhttp.nonProxyHosts` argument as this arg
may be spilled over to an argsfile, where quoting is considered by the
JVM as part of the host expression.  (I expect the single-quoting to be
unnecessary in the pre-Java9+ path, as my understanding of `rctx.execute`
is that Bazel _doesn't_ use the shell, so special chars like pipes, etc.
are harmless.)

Resolves #780.