Remove unused test
This test has been failing for a while, and yet we've never
noticed, and we don't seem to care. It depends on us exposing
the cached files from coursier. Given that this code is stable
and unlikely to change (see https://github.com/bazelbuild/rules_jvm_external/pull/807)
this feels like a safe change to make.