Deprecate jetify
AndroidX is a series of dependencies that Google released, like standard libraries but
for all things Google. At some point they switched the namespace on those libraries to
`androidx`, which caused a ton of runtime crashes because you'd have older versions of
libraries that hadn't updated to Androidx conflicting with ones that have updated.

The last release of jetifier was in 2020, and since then the tool has been deprecated
& so we should also mark it as deprecated and remove from rules_jvm_external a future
release.