Prepare for Java 8 support dropping from modern JDKs
Modern JDKs print a warning that Java 8 support is deprecated and
will be removed in a future release. To prevent that warning being
displayed, we add an additional `javac` flag when compiling the
code which targets Java 8.

The long-term solution is to only target Java 11+, which I suggest
we do after Java 21 ships in Septembe after 2023.