Introduce a maven-based resolver
This resolver uses the standard Maven resolver under the covers, but also
allows us to offer support for using Maven BOMs during artifact resolution
which will resolve a long-standing issue.

To avoid requiring a prebuilt-binary, this rule only allows us to pin
dependencies when a lock file is being used.