maven_install: fail if 'repositories' attribute is empty
Coursier fails with a cryptic error in case someone tries to use
`maven_install()` without specifying any `repositories`, which is a
small silly thing I did when trying to extract internal corporate code
into a public proof-of-concept on how to do something with Bazel that
took me 10 minutes to hunt down :)

When `repositories` is empty, Coursier fails with an error like:

```
  Dependencies:
org.junit.jupiter:junit-jupiter-engine:5.10.2:
Resolution error: Error downloading org.junit.jupiter:junit-jupiter-engine:5.10.2
```

this was for a `maven_install()` block with a single artifact to install;
otherwise the error message will contain one `Resolution error` per artifact.
The problem though is that Coursier doesn't say what the error is, just that
one occurred.

The root-cause is that `coursier_fetch` etc will invoke coursier with arguments like

```
coursier fetch <artifacts> ... --no-default ...
```

So as a simple fix to prevent people from configuring `maven_install` in a way
that doesn't make sense like I did, this commit adds a check inside
`make_coursier_dep_tree` (where the arguments to coursier are constructed) to
fail if the list/dict of repositories is empty.