handle transitive dependencies of packaging=pom
closes #1206

Maven interprets a `<dependency>` (not in the `<dependencyManagement>`
section) with `type=pom` as adding all of those artifact's dependencies
(direct and transitive) to the current module's list of dependencies.

This commit modifies RJE to replicate that behavior, so that any
artifact which is directly referenced in `maven_install()` also has all
of its true transitive dependencies imported/installed into the Bazel
workspace.

The fix here is to append `pom` to the list of types passed to
coursier's `--artifact-types` flag so that it properly fetches/resolves
the type=pom artifact, and to append a condition to the if-else branch
meant to support direct dependencies with type=pom in `maven_install()`
(see #99) to handle a transitive dependency on type=pom as well.