Remove the Windows `kt_jvm_export` example (#1383)
This has been failing for some time, and we've never understood why. Since none
of the committers on the project have access to a Windows machine, and this is
meant to be an example, it seems better to disable this build rather than have
unfixable errors occur on people's PRs.